+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Last Changes
 
+- [#6](https://github.com/alex3683/hal-http-client/issues/6): added missing body JSON stringify
+
+
 ## v0.4.0
 
 - [#5](https://github.com/alex3683/hal-http-client/issues/5): using correct body field